Internal change
PiperOrigin-RevId: 453491226
