+++ base/BUILD
@@ -103,35 +103,89 @@ cc_library(
     name = "type",
     srcs = [
         "type.cc",
-        "type_factory.cc",
-        "type_manager.cc",
-        "type_provider.cc",
     ],
     hdrs = [
         "type.h",
-        "type_factory.h",
-        "type_manager.h",
-        "type_provider.h",
-        "type_registry.h",
     ],
     deps = [
         ":handle",
         ":kind",
         ":memory_manager",
         "//base/internal:type",
-        "//internal:casts",
-        "//internal:no_destructor",
-        "//internal:rtti",
+        "@com_google_absl//absl/hash",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/types:variant",
+    ],
+)
+
+cc_library(
+    name = "type_manager",
+    srcs = ["type_manager.cc"],
+    hdrs = ["type_manager.h"],
+    deps = [
+        ":type",
+        ":type_factory",
+        ":type_provider",
         "//internal:status_macros",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/container:flat_hash_map",
-        "@com_google_absl//absl/hash",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/synchronization",
+    ],
+)
+
+cc_library(
+    name = "type_provider",
+    srcs = ["type_provider.cc"],
+    hdrs = ["type_provider.h"],
+    deps = [
+        ":handle",
+        ":type",
+        ":type_factory",
+        "//internal:no_destructor",
+        "@com_google_absl//absl/base:core_headers",
+        "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/status:statusor",
         "@com_google_absl//absl/strings",
+    ],
+)
+
+cc_library(
+    name = "type_registry",
+    hdrs = ["type_registry.h"],
+    deps = [":type_provider"],
+)
+
+cc_library(
+    name = "type_factory",
+    srcs = ["type_factory.cc"],
+    hdrs = ["type_factory.h"],
+    deps = [
+        ":handle",
+        ":memory_manager",
+        "//base/types:any",
+        "//base/types:bool",
+        "//base/types:bytes",
+        "//base/types:double",
+        "//base/types:duration",
+        "//base/types:dyn",
+        "//base/types:enum",
+        "//base/types:error",
+        "//base/types:int",
+        "//base/types:list",
+        "//base/types:map",
+        "//base/types:null",
+        "//base/types:string",
+        "//base/types:struct",
+        "//base/types:timestamp",
+        "//base/types:type",
+        "//base/types:uint",
+        "@com_google_absl//absl/base:core_headers",
+        "@com_google_absl//absl/container:flat_hash_map",
+        "@com_google_absl//absl/status",
         "@com_google_absl//absl/synchronization",
-        "@com_google_absl//absl/types:span",
-        "@com_google_absl//absl/types:variant",
     ],
 )
 
@@ -145,6 +199,8 @@ cc_test(
         ":handle",
         ":memory_manager",
         ":type",
+        ":type_factory",
+        ":type_manager",
         ":value",
         "//base/internal:memory_manager_testing",
         "//internal:testing",
@@ -169,7 +225,25 @@ cc_library(
         ":kind",
         ":memory_manager",
         ":type",
+        ":type_manager",
         "//base/internal:value",
+        "//base/types:any",
+        "//base/types:bool",
+        "//base/types:bytes",
+        "//base/types:double",
+        "//base/types:duration",
+        "//base/types:dyn",
+        "//base/types:enum",
+        "//base/types:error",
+        "//base/types:int",
+        "//base/types:list",
+        "//base/types:map",
+        "//base/types:null",
+        "//base/types:string",
+        "//base/types:struct",
+        "//base/types:timestamp",
+        "//base/types:type",
+        "//base/types:uint",
         "//internal:casts",
         "//internal:no_destructor",
         "//internal:rtti",
@@ -201,6 +275,8 @@ cc_test(
     deps = [
         ":memory_manager",
         ":type",
+        ":type_factory",
+        ":type_manager",
         ":value",
         "//base/internal:memory_manager_testing",
         "//internal:strings",
