+++ base/type_factory.h
@@ -24,7 +24,23 @@
 #include "absl/synchronization/mutex.h"
 #include "base/handle.h"
 #include "base/memory_manager.h"
-#include "base/type.h"
+#include "base/types/any_type.h"
+#include "base/types/bool_type.h"
+#include "base/types/bytes_type.h"
+#include "base/types/double_type.h"
+#include "base/types/duration_type.h"
+#include "base/types/dyn_type.h"
+#include "base/types/enum_type.h"
+#include "base/types/error_type.h"
+#include "base/types/int_type.h"
+#include "base/types/list_type.h"
+#include "base/types/map_type.h"
+#include "base/types/null_type.h"
+#include "base/types/string_type.h"
+#include "base/types/struct_type.h"
+#include "base/types/timestamp_type.h"
+#include "base/types/type_type.h"
+#include "base/types/uint_type.h"
 
 namespace cel {
