+++ base/type_test.cc
@@ -204,7 +204,7 @@ TEST_P(TypeTest, MoveConstructor) {
   Persistent<const Type> from(type_factory.GetIntType());
   Persistent<const Type> to(std::move(from));
   IS_INITIALIZED(from);
-  EXPECT_EQ(from, type_factory.GetNullType());
+  EXPECT_FALSE(from);
   EXPECT_EQ(to, type_factory.GetIntType());
 }
 
@@ -221,7 +221,7 @@ TEST_P(TypeTest, MoveAssignment) {
   Persistent<const Type> to(type_factory.GetNullType());
   to = std::move(from);
   IS_INITIALIZED(from);
-  EXPECT_EQ(from, type_factory.GetNullType());
+  EXPECT_FALSE(from);
   EXPECT_EQ(to, type_factory.GetIntType());
 }
