+++ base/types/enum_type.cc
@@ -1,0 +1,41 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "base/types/enum_type.h"
+
+#include "absl/status/statusor.h"
+#include "absl/strings/string_view.h"
+#include "absl/types/variant.h"
+
+namespace cel {
+
+CEL_INTERNAL_TYPE_IMPL(EnumType);
+
+struct EnumType::FindConstantVisitor final {
+  const EnumType& enum_type;
+
+  absl::StatusOr<Constant> operator()(absl::string_view name) const {
+    return enum_type.FindConstantByName(name);
+  }
+
+  absl::StatusOr<Constant> operator()(int64_t number) const {
+    return enum_type.FindConstantByNumber(number);
+  }
+};
+
+absl::StatusOr<EnumType::Constant> EnumType::FindConstant(ConstantId id) const {
+  return absl::visit(FindConstantVisitor{*this}, id.data_);
+}
+
+}  // namespace cel
